CAN/c.sharp: don't cast IntPtr to int
Fixes the wrapper usage on 64-bit systems.

+++ CAN/c.sharp/vs_can_api_wrapper.cs
@@ -449,7 +449,7 @@ namespace VSCom.CanApi
 			for (int i = 0; i < Read; i++) {
 				Buf[i] = (VSCAN_MSG)Marshal.PtrToStructure(pCurrent, typeof(VSCAN_MSG));
 
-				pCurrent = (IntPtr)((int)pCurrent + Marshal.SizeOf(Buf[0]));
+				pCurrent = pCurrent + Marshal.SizeOf(Buf[0]);
 			}
 
 			Marshal.FreeHGlobal(pBuf);
